Move estop to plugins folder
