+++ panther_gazebo/src/plugins/Estop.qrc
